Finished the data decoding block and implemented both dew point and RealFeel© calculation

+++ data_decoder.h
@@ -6,6 +6,6 @@ std::list<float> decode_humidity (unsigned char* raw_data);
 float decode_wind_chill (unsigned char* raw_data);
 float decode_wind_gust (unsigned char* raw_data);
 float decode_wind_speed (unsigned char* raw_data);
-int decode_wind_dir (unsigned char* raw_data);
+float decode_wind_dir (unsigned char* raw_data);
 
 int bcd2int(char bcd);
